# specs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API enables end-users, owners and Sally devices to interact  with the central database.",
    "version": "1.0",
    "title": "Pantry REST API",
    "termsOfService": "http://casabots.com/terms/",
    "contact": {},
    "license": {
      "name": "Copyright (c) CASABOTS 2016"
    }
  },
  "host": "localhost:8080",
  "basePath": "/pantry",
...
  "paths": {
...
    "/ingredients": {
      "post": {
        "tags": [
          "ingredient-controller"
        ],
        "summary": "createIngredient",
        "operationId": "createIngredientUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "ingredient",
            "description": "ingredient",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Ingredient"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Ingredient"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/ingredients/update": {
      "post": {
        "tags": [
          "ingredient-controller"
        ],
        "summary": "updateIngredient",
        "operationId": "updateIngredientUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "ingredient",
            "description": "ingredient",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Ingredient"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Ingredient"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/ingredients/{ingredientId}": {
      "get": {
        "tags": [
          "ingredient-controller"
        ],
        "summary": "getIngredient",
        "operationId": "getIngredientUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "ingredientId",
            "in": "path",
            "description": "ingredientId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Ingredient"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "ingredient-controller"
        ],
        "summary": "deleteIngredient",
        "operationId": "deleteIngredientUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "ingredientId",
            "in": "path",
            "description": "ingredientId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "204": {
            "description": "No Content"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Ingredient": {
      "type": "object",
      "properties": {
        "available": {
          "type": "boolean"
        },
        "category": {
          "type": "string"
        },
        "imageUrl": {
          "type": "string"
        },
        "ingredientId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
...
  }
}